🚚 move CommandArgument into their own package
