+++ src/main/java/azerty/tguichaoua/mpb/command/argument/GlobListedCommandArgument.java
@@ -1,5 +1,7 @@
-package azerty.tguichaoua.mpb.command;
+package azerty.tguichaoua.mpb.command.argument;
 
+import azerty.tguichaoua.mpb.command.CommandException;
+import azerty.tguichaoua.mpb.command.CommandExecution;
 import azerty.tguichaoua.mpb.util.RegexUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
